\name{zooimage-package}
\alias{zooimage-package}
\alias{zooimage}
\docType{package}

\title{ZooImage: analysis of digital plankton images}
\description{
  ZooImage is a free (open source) solution for analyzing digital
  images of plankton. In combination with ImageJ, a free image analysis
  system, it processes digital images, measures individuals, trains for
  automatic classification of taxa, and finally, measures plankton samples
  (abundances, total and partial size spectra or biomasses, etc.)
}

\details{
  \tabular{ll}{
    Package: \tab zooimage\cr
    Type: \tab Package\cr
    Version: \tab 5.4-3\cr
    Date: \tab 2015-10-09\cr
    License: \tab GPL 2 or above at your convenience.\cr
  }
}

\author{
  Philippe Grosjean, Kevin Denis & Guillaume Wacquet, Numerical Ecology of
  Aquatic Systems, Mons University, Belgium.
  A part of the early concepts of ZooImage are inspired from PVA by Xabier Irigoien,
  Guillermo Boyra & Igor Arregi, AZTI Technalia, Spain, with their autorization.
  Angel Lopez-Urrutia, Centro Oceanografico de Gijon, IEO, Spain, also contributed
  to the initial concept during early discussions. No code come from them, however.
  Mike Sieracki, Ben Tupper et al for the FIT VIS plugin (FlowCAM images
  process in ImageJ -FitVIS-, but this pulgin is not used any more by default in
  the current version).

  Maintainer: Philippe Grosjean <Philippe.Grosjean@umons.ac.be>
  
  The software was funded, and is the property of:
  
  - The University of Mons, Belgium,
  - Belgian Scientific Policy (BelSpo), Belgium, that contributed by funding the
    AMORE III project.
  - IFREMER, France, through a co-funded research collaboration with the
    University of Mons.
}

%\references{
%  Not yet!
%}

%\examples{
%# TODO...
%}

\keyword{package}
