\name{zooimage-package}
\alias{zooimage-package}
\alias{zooimage}
\docType{package}

\title{ZooImage: analysis of digital plankton images}
\description{
  ZooImage is a free (open source) solution for analyzing digital
  images of plankton. In combination with ImageJ, a free image analysis
  system, it processes digital images, measures individuals, trains for
  automatic identification of taxa, and finally, measures plankton samples
  (abundances, total and partial size spectra or biomasses, etc.)
}

\details{
  \tabular{ll}{
    Package: \tab zooimage\cr
    Type: \tab Package\cr
    Version: \tab 2.0-0\cr
    Date: \tab 2010-04-06\cr
    License: \tab GPL 2 or above at your convenience.\cr
  }
  For any series that you analyses with ZooImage, could you, please:
  - Send main metadata and a short description of your series, and the way it
    is analyzed to Philippe.Grosjean@umons.ac.be (this is only used internally
    in order to keep track of the way ZooImage is used).
  - Everytime you publish results that use ZooImage, you must place a reference
    to the ZooImage web site (http://www.sciviews.org/zooimage) in your
	publication. For papers, please, send also a reprint to
	Philippe.Grosjean@umons.ac.be, preferrably as a PDF file.
}

\author{
  Philippe Grosjean & Kevin Denis, Numerical Ecology of Aquatic Systems,
  Mons University, Belgium.
  Xabier Irigoien, Guillermo Boyra & Igor Arregi, AZTI Technalia, Spain (PVA,
  whose a part of ZooImage is inspired from).
  Angel Lopez-Urrutia, Centro Oceanografico de Gijon, IEO, Spain (initial idea).
  Mike Sieracki, Ben Tupper et al for the FIT VIS plugin (FlowCAM images process).

  Maintainer: Philippe Grosjean <Philippe.Grosjean@umons.ac.be>
}

\references{
  Not yet!
}

\examples{
# TODO...
}

\keyword{package}